Added ccxt library Moved MarketWatcher to financial module
