+++ requirements.txt
@@ -1,9 +1,17 @@
+aiodns==1.1.1
+aiohttp==2.3.7
+async-timeout==2.0.0
+cchardet==2.1.1
+ccxt==1.10.544
 certifi==2017.11.5
 chardet==3.0.4
 colorama==0.3.9
 idna==2.6
+multidict==3.3.2
 numpy==1.13.3
+pycares==2.3.0
 pymongo==3.6.0
 requests==2.18.4
 TA-Lib==0.4.10
 urllib3==1.22
+yarl==0.17.0
